+++ tests/core/incubation/test_command.py
@@ -16,11 +16,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """Test command builder."""
-import threading
 
 from renku.core.incubation.command import Command, CommandResult
-from renku.core.utils import communication
-from renku.service.utils.communication import ServiceCallback
 
 
 def test_dataset_add_command(project, tmp_path):
@@ -49,51 +46,3 @@ def test_dataset_add_command(project, tmp_path):
 
     assert isinstance(files, list)
     assert "some-file" in [file_.name for file_ in files]
-
-
-def test_communicator_is_unsubscribed(project, tmp_path):
-    """Test communicator is unsubscribed when command is executed."""
-    communicator = ServiceCallback()
-
-    command = Command().command(lambda _: communication.echo("Hello world!"))
-    command.with_communicator(communicator).build().execute()
-
-    assert ["Hello world!"] == communicator.messages
-
-    communication.echo("More messages.")
-
-    assert ["Hello world!"] == communicator.messages
-
-
-def test_multi_communicators(project, tmp_path):
-    """Test subscribing multiple communicators."""
-    communicator_1 = ServiceCallback()
-    communicator_2 = ServiceCallback()
-
-    command = Command().command(lambda _: communication.echo("Hello world!"))
-    command.with_communicator(communicator_1).with_communicator(communicator_2).build().execute()
-
-    assert ["Hello world!"] == communicator_1.messages
-    assert ["Hello world!"] == communicator_2.messages
-
-
-def test_multi_threaded_communication(project, tmp_path):
-    """Test communication with multi-threading."""
-
-    def thread_function(name, communicator):
-        command = Command().command(lambda _: communication.echo(f"Hello world from {name}!"))
-        command.with_communicator(communicator).build().execute()
-
-    communicator_1 = ServiceCallback()
-    thread_1 = threading.Thread(target=thread_function, args=("thread-1", communicator_1))
-
-    communicator_2 = ServiceCallback()
-    thread_2 = threading.Thread(target=thread_function, args=("thread-2", communicator_2))
-
-    thread_1.start()
-    thread_2.start()
-    thread_1.join()
-    thread_2.join()
-
-    assert ["Hello world from thread-1!"] == communicator_1.messages
-    assert ["Hello world from thread-2!"] == communicator_2.messages
